Fix directory copying

+++ src/linker.rs
@@ -208,15 +208,75 @@ impl Linker {
                 })?;
             }
             LinkType::Copy => {
-                fs::copy(&src, &dest).with_context(|| {
-                    format!(
-                        "Failed to copy from {} to {}",
-                        src.as_ref().display(),
-                        dest.as_ref().display()
-                    )
-                })?;
+                if src.as_ref().is_file() {
+                    fs::copy(&src, &dest).with_context(|| {
+                        format!(
+                            "Failed to copy from {} to {}",
+                            src.as_ref().display(),
+                            dest.as_ref().display()
+                        )
+                    })?;
+                } else if src.as_ref().is_dir() {
+                    self.copy_dir(&src, &dest).with_context(|| {
+                        format!(
+                            "Failed to copy dir from {} to {}",
+                            src.as_ref().display(),
+                            dest.as_ref().display()
+                        )
+                    })?;
+                } else {
+                    return Err(anyhow!("Cannot copy from path: {}", src.as_ref().display()));
+                }
             }
         };
+
+        Ok(())
+    }
+
+    pub fn copy_dir<U: AsRef<Path>, V: AsRef<Path>>(
+        &self,
+        from: U,
+        to: V,
+    ) -> Result<(), std::io::Error> {
+        let mut stack = Vec::new();
+        stack.push(PathBuf::from(from.as_ref()));
+
+        let output_root = PathBuf::from(to.as_ref());
+        let input_root = PathBuf::from(from.as_ref()).components().count();
+
+        while let Some(working_path) = stack.pop() {
+            // Generate a relative path
+            let src: PathBuf = working_path.components().skip(input_root).collect();
+
+            // Create a destination if missing
+            let dest = if src.components().count() == 0 {
+                output_root.clone()
+            } else {
+                output_root.join(&src)
+            };
+
+            if fs::metadata(&dest).is_err() {
+                fs::create_dir_all(&dest)?;
+            }
+
+            for entry in fs::read_dir(working_path)? {
+                let entry = entry?;
+                let path = entry.path();
+                if path.is_dir() {
+                    stack.push(path);
+                } else {
+                    match path.file_name() {
+                        Some(filename) => {
+                            let dest_path = dest.join(filename);
+                            fs::copy(&path, &dest_path)?;
+                        }
+                        None => {
+                            println!("failed: {:?}", path);
+                        }
+                    }
+                }
+            }
+        }
 
         Ok(())
     }
